+++ typescript/metrics/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @hyperlane-xyz/metrics
 
+## 0.2.32
+
+### Patch Changes
+
+- Updated dependencies [4976bb1]
+- Updated dependencies [6c9210b]
+- Updated dependencies [086ec59]
+  - @hyperlane-xyz/sdk@39.1.0
+  - @hyperlane-xyz/utils@39.1.0
+  - @hyperlane-xyz/core@11.3.1
+
 ## 0.2.31
 
 ### Patch Changes